Refactor schedule and event details generation in common module
