+++ functions/src/main.ts
@@ -2,10 +2,10 @@ import { initializeApp } from 'firebase-admin'
 import { config, https } from 'firebase-functions'
 import fetch from 'node-fetch'
 
-import { generateEventDetails } from './event-details-view/generate-event-details'
+import { generateEventDetails } from './events-view/generate-event-details'
+import { generateSchedule } from './events-view/generate-schedule'
 import { migrateToFirestore } from './migrateToFirestore'
 import { patch } from './patch/patch'
-import { generateSchedule } from './schedule-view/generate-schedule'
 import { indexContent } from './index-contents/index-contents'
 import { generateSpeakers } from './speakers-view/generate-speakers'
 import { generateTracks } from './tracks-view/generate-tracks'
