--- functions/src/schedule-view/schedule-view-data.ts
@@ -1,15 +1,0 @@
-import { Event } from '../event-details-view/event-details-view-data'
-
-export interface Schedule {
-    pages: SchedulePage[]
-}
-
-export interface SchedulePage {
-    day: Day
-    events: Event[]
-}
-
-export interface Day {
-    id: string
-    date: Date
-}
